analytics-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
